requirements.txt
tweepy~=3.10.0
PyYAML~=5.4.1